Moved Rating Component logic out of Game Overlay to Rating Component
