+++ client/components/GameOverlay.jsx
@@ -4,25 +4,12 @@ import RatingComponent from './RatingComponent'
 import Scoreboard from './Scoreboard'
 import Scoreboard_v2 from './Scoreboard_v2'
 
-import { getById } from '../utility/api-tools.js'
 /**
  * This is the page where you can play the game and see your score on the side.
  */
 
 const GameOverlay = ({ isPlaying, closeOverlay, gameid, scores }) => {
   // try to refactor in the old game page grid layout
-  const [positivePercentage, setPositivePercentage] = useState(0)
-
-  useEffect(() => {
-    if(gameid !== 0) {
-      getById('/game/getOverallRating', gameid, setPercentage)
-    }
-  }, [isPlaying])
-
-  const setPercentage = (response) => {
-    console.log(response.positivePercent)
-    setPositivePercentage(response.positivePercent)
-  }
 
   return (
     <div className={'game-overlay'}
@@ -38,7 +25,8 @@ const GameOverlay = ({ isPlaying, closeOverlay, gameid, scores }) => {
         <div className="item3"><GameComponent gameid={gameid} /></div>
         <div className="item4">
           <RatingComponent
-            positivePercentage={positivePercentage}
+            isPlaying={isPlaying}
+            gameid={gameid}
           />
         </div>
         <div className="item5"></div>
